+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -13,21 +13,20 @@ import 'instance_manager.dart';
 /// Converts [WebResourceRequestData] to [WebResourceRequest]
 WebResourceRequest _toWebResourceRequest(WebResourceRequestData data) {
   return WebResourceRequest(
-    url: data.url!,
-    isForMainFrame: data.isForMainFrame!,
+    url: data.url,
+    isForMainFrame: data.isForMainFrame,
     isRedirect: data.isRedirect,
-    hasGesture: data.hasGesture!,
-    method: data.method!,
-    requestHeaders:
-        data.requestHeaders?.cast<String, String>() ?? <String, String>{},
+    hasGesture: data.hasGesture,
+    method: data.method,
+    requestHeaders: data.requestHeaders.cast<String, String>(),
   );
 }
 
 /// Converts [WebResourceErrorData] to [WebResourceError].
 WebResourceError _toWebResourceError(WebResourceErrorData data) {
   return WebResourceError(
-    errorCode: data.errorCode!,
-    description: data.description!,
+    errorCode: data.errorCode,
+    description: data.description,
   );
 }
 
@@ -115,8 +114,8 @@ class WebViewHostApiImpl extends WebViewHostApi {
   Future<void> loadDataFromInstance(
     WebView instance,
     String data,
-    String mimeType,
-    String encoding,
+    String? mimeType,
+    String? encoding,
   ) {
     return loadData(
       instanceManager.getInstanceId(instance)!,
@@ -129,11 +128,11 @@ class WebViewHostApiImpl extends WebViewHostApi {
   /// Helper method to convert instances ids to objects.
   Future<void> loadDataWithBaseUrlFromInstance(
     WebView instance,
-    String baseUrl,
+    String? baseUrl,
     String data,
-    String mimeType,
-    String encoding,
-    String historyUrl,
+    String? mimeType,
+    String? encoding,
+    String? historyUrl,
   ) {
     return loadDataWithBaseUrl(
       instanceManager.getInstanceId(instance)!,
@@ -164,7 +163,7 @@ class WebViewHostApiImpl extends WebViewHostApi {
   }
 
   /// Helper method to convert instances ids to objects.
-  Future<String> getUrlFromInstance(WebView instance) {
+  Future<String?> getUrlFromInstance(WebView instance) {
     return getUrl(instanceManager.getInstanceId(instance)!);
   }
 
@@ -202,16 +201,18 @@ class WebViewHostApiImpl extends WebViewHostApi {
   }
 
   /// Helper method to convert instances ids to objects.
-  Future<String> evaluateJavascriptFromInstance(
+  Future<String?> evaluateJavascriptFromInstance(
     WebView instance,
     String javascriptString,
   ) {
     return evaluateJavascript(
-        instanceManager.getInstanceId(instance)!, javascriptString);
+      instanceManager.getInstanceId(instance)!,
+      javascriptString,
+    );
   }
 
   /// Helper method to convert instances ids to objects.
-  Future<String> getTitleFromInstance(WebView instance) {
+  Future<String?> getTitleFromInstance(WebView instance) {
     return getTitle(instanceManager.getInstanceId(instance)!);
   }